+++ cmd/tools/root.go
@@ -6,6 +6,8 @@ import (
 	"github.com/dolittle/platform-api/cmd/tools/job"
 	"github.com/dolittle/platform-api/cmd/tools/studio"
 	"github.com/dolittle/platform-api/cmd/tools/terraform"
+
+	"github.com/dolittle/platform-api/cmd/tools/users"
 	"github.com/spf13/cobra"
 )
 
@@ -23,4 +25,5 @@ func init() {
 	RootCmd.AddCommand(explore.RootCmd)
 	RootCmd.AddCommand(job.RootCMD)
 	RootCmd.AddCommand(terraform.RootCMD)
+	RootCmd.AddCommand(users.RootCMD)
 }